refactor: move extension source under src
Keep the package entry point in src/index.ts and separate implementation modules from tests and repository metadata.
